+++ docs/developer_guide.rst
@@ -2,6 +2,6 @@
 Developers
 **********
 
-This guide provides instructions for setting up `mast-aladin-lite` for development.
+This guide provides instructions for setting up `mast-aladin` for development.
 
 See the **User Installation** section for installation steps.
