+++ licenses/README.rst
@@ -1,8 +1,8 @@
 Licenses
 ========
 
-This directory holds license and credit information for works the mast_aladin_lite
+This directory holds license and credit information for works the mast_aladin
 package is derived from or distributes, and/or datasets.
 
-The license file for the mast_aladin_lite package itself is located in the root of
+The license file for the mast_aladin package itself is located in the root of
 this repository.
