+++ mast_aladin/__init__.py
@@ -1,0 +1,7 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Top-level package for mast_aladin."""
+
+# Top-level API as exposed to users.
+from mast_aladin.app import MastAladin, gca  # noqa: F401
+from mast_aladin.table import MastTable  # noqa: F401
+from mast_aladin.app_sidecar import AppSidecar  # noqa: F401
