+++ mast_aladin/app.py
@@ -1,9 +1,9 @@
 from ipyaladin import Aladin
-from mast_aladin_lite.aida import AID
-from mast_aladin_lite.table import MastTable
-from mast_aladin_lite.mixins import DelayUntilRendered
-from mast_aladin_lite.overlay.overlay_manager import OverlayManager
-from mast_aladin_lite.overlay.mast_overlay import MastOverlay
+from mast_aladin.aida import AID
+from mast_aladin.table import MastTable
+from mast_aladin.mixins import DelayUntilRendered
+from mast_aladin.overlay.overlay_manager import OverlayManager
+from mast_aladin.overlay.mast_overlay import MastOverlay
 import io
 from ipyaladin.elements.error_shape import (
     CircleError,
@@ -315,7 +315,7 @@ class MastAladin(Aladin, DelayUntilRendered):
 
 def gca():
     """
-    Get the current mast-aladin-lite application instance.
+    Get the current mast-aladin application instance.
     If none exist, create a new one.
 
     Returns
