+++ mast_aladin/table/data/column_descriptions.json
