+++ mast_aladin/table/mast_table.vue
