+++ mast_aladin/table/validate.py
@@ -90,7 +90,7 @@ def update_mast_column_lists(update_column_descriptions=True, update_unique_colu
 
     To update the cached columns, run:
 
-        from mast_aladin_lite.table import validate
+        from mast_aladin.table import validate
         validate.update_mast_column_lists()
     """
     from astroquery.mast.missions import MastMissions
