+++ mast_aladin/tests/test_aida_example.py
