+++ mast_aladin/tests/test_example.py
@@ -1,4 +1,4 @@
-from mast_aladin_lite import MastAladin
+from mast_aladin import MastAladin
 
 
 def test_instance_creation(MastAladin_app):
