+++ mast_aladin/tests/test_overlays_dict.py
@@ -6,8 +6,8 @@ from unittest.mock import Mock
 import warnings
 import re
 
-from mast_aladin_lite import MastAladin
-from mast_aladin_lite.overlay.mast_overlay import MastOverlay, MastOverlayType
+from mast_aladin import MastAladin
+from mast_aladin.overlay.mast_overlay import MastOverlay, MastOverlayType
 from ipyaladin.elements.error_shape import EllipseError, CircleError
 from regions import CircleSkyRegion
 from ipyaladin import Marker
