+++ mast_aladin/tests/test_overlays_update.py
@@ -1,8 +1,8 @@
 from astropy.table import Table
 from astropy.coordinates import SkyCoord, Angle
 
-from mast_aladin_lite import MastAladin
-from mast_aladin_lite.overlay.mast_overlay import MastOverlay
+from mast_aladin import MastAladin
+from mast_aladin.overlay.mast_overlay import MastOverlay
 from ipyaladin.elements.error_shape import CircleError
 from regions import CircleSkyRegion
 from ipyaladin import Marker
