--- mast_aladin_lite/__init__.py
@@ -1,7 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Top-level package for mast_aladin_lite."""
-
-# Top-level API as exposed to users.
-from mast_aladin_lite.app import MastAladin, gca  # noqa: F401
-from mast_aladin_lite.table import MastTable  # noqa: F401
-from mast_aladin_lite.app_sidecar import AppSidecar  # noqa: F401
