+++ tox.ini
@@ -12,7 +12,7 @@ requires =
 isolated_build = true
 
 [testenv]
-description = Run tests for mast-aladin-lite
+description = Run tests for mast-aladin
 # Suppress display of matplotlib plots generated during docs build
 setenv =
     MPLBACKEND=agg
@@ -47,8 +47,8 @@ extras =
 commands =
     #jupyter --paths
     pip freeze
-    !cov: pytest --pyargs mast_aladin_lite {toxinidir}/docs
-    cov: pytest --pyargs mast_aladin_lite {toxinidir}/docs --cov mast_aladin_lite --cov-config={toxinidir}/pyproject.toml
+    !cov: pytest --pyargs mast_aladin {toxinidir}/docs
+    cov: pytest --pyargs mast_aladin {toxinidir}/docs --cov mast_aladin --cov-config={toxinidir}/pyproject.toml
     cov: coverage xml -o {toxinidir}/coverage.xml
 
 [testenv:linkcheck]
@@ -64,14 +64,14 @@ skip_install = true
 changedir = .
 description = check code style, e.g. with flake8
 deps = flake8
-commands = flake8 mast_aladin_lite --count
+commands = flake8 mast_aladin --count
 
 [testenv:securityaudit]
 skip_install = true
 changedir = .
 description = security audit with bandit
 deps = bandit
-commands = bandit -r mast_aladin_lite -c .bandit.yaml
+commands = bandit -r mast_aladin -c .bandit.yaml
 
 [testenv:pep517]
 skip_install = true
